Reorganized polluted location components
