+++ src/components/pollutedLocationsBar/PollutedLocationCard.tsx
