+++ src/components/pollutedLocationsBar/PollutedLocationFormResult.tsx
