+++ src/components/pollutedLocationsBar/PollutedLocationList.tsx
